--- coffeescript-symbols/@types/coffeescript/index.d.ts
@@ -1,4 +1,0 @@
-declare module "coffeescript" {
-  import { Block } from "coffeescript/lib/coffeescript/nodes"
-  export function nodes(source: string, options?: any): Block;
-}
