--- coffeescript-symbols/@types/coffeescript/scope.d.ts
@@ -1,7 +1,0 @@
-declare module "coffeescript/lib/coffeescript/scope" {
-  export class Scope {
-    variables: { name: string, type: string}[]
-    positions: any
-    utilities?: any
-  }
-}
