+++ coffeescript-symbols/package.json
@@ -17,6 +17,7 @@
   },
   "devDependencies": {
     "@types/jest": "^23.3.1",
+    "coffeescript-types": "file:../coffeescript-types",
     "jest": "^23.5.0",
     "ts-jest": "^23.1.4"
   }
