+++ coffeescript-types/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "coffeescript-types",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "types": "index.d.ts",
+  "author": "",
+  "license": "ISC"
+}
